[Workflow] Release Notes => Change Log
The focus of these changes is to update the release process to explicitly point the "Release Notes" on the GH release page to the change log for the release.  This will ensure that default content is not generated, and we have one consistent source of truth for the changes in a release.