Add `ResponseItem` constructors (#625)
All the classes derived from `ResponseItem` now have public constructors, and these constructors are generated.

A side effect of making these classes behave as both, inputs and outputs, was that their required properties got setters. Setters make sense when the values of these properties can be updated, but in this particular case they should only be settable at construction time. This is a limitation in the code generator being tracked by the following issue:
🔗 https://github.com/microsoft/typespec/issues/5175#issuecomment-3222039076