Make id required for McpToolCallApprovalRequestItem (#671)
Replaced the constructor that does not take the item ID as a parameter in favor of one that does. This is because MCP approval requests are correlated to MCP approval responses using this ID. Therefore, it implies that the ID of MCP approval requests is required. Note that this is contrary to other similar cases, such as function call items, which are instead correlated to function call outputs using a dedicated "call ID".